requirements.txt
python-dotenv
torch
llama-index
llama-index-llms-huggingface
